docs(examples): simplify vitest examples
