+++ examples/zimic-interceptor-with-openapi-typegen/package.json
@@ -17,7 +17,6 @@
   },
   "devDependencies": {
     "@types/supertest": "^6.0.3",
-    "@vitest/spy": "^3.1.1",
     "@zimic/interceptor": "latest",
     "dotenv-cli": "^8.0.0",
     "supertest": "^7.1.0",
