+++ examples/zimic-interceptor-with-vitest-browser/README.md
@@ -4,8 +4,7 @@
 
 This example uses [@zimic/interceptor](https://www.npmjs.com/package/@zimic/interceptor) with
 [Vitest](https://vitest.dev) with [Browser Mode](https://vitest.dev/guide/browser) enabled. It uses
-[Playwright](https://playwright.dev) as the browser provider for Vitest and
-[Testing Library](https://testing-library.com).
+[Playwright](https://playwright.dev) as the browser provider for Vitest.
 
 - [Application](#application)
 - [Testing](#testing)
@@ -15,19 +14,13 @@ This example uses [@zimic/interceptor](https://www.npmjs.com/package/@zimic/inte
 
 ## Application
 
-The application is a simple HTML layout rendered by vanilla JavaScript, fetching repositories from the
-[GitHub API](https://docs.github.com/en/rest).
-
-- Application: [`src/app.ts`](./src/app.ts)
-
 A `postinstall` script in [`package.json`](./package.json) is used to install Playwright's browsers and initialize
 Zimic's mock service worker to the `./public` directory. The mock service worker at `./public/mockServiceWorker.js` is
 ignored in the [`.gitignore`](./.gitignore) file.
 
 ## Testing
 
-An example test suite uses Vitest to test the application. `@zimic/interceptor` is used to mock the GitHub API and
-simulate a test case where the repository is found and another where it is not.
+`@zimic/interceptor` is used to mock requests made to the [GitHub API](https://docs.github.com/rest).
 
 ### `@zimic/interceptor`
