+++ examples/zimic-interceptor-with-vitest-browser/tests/example.test.ts
@@ -1,52 +1,32 @@
-import { screen } from '@testing-library/dom';
-import userEvent from '@testing-library/user-event';
-import { beforeEach, describe, expect, it } from 'vitest';
+import { describe, expect, it } from 'vitest';
 
-import renderApp from '../src/app';
-import { GitHubRepository } from '../src/clients/github';
+import { GITHUB_API_BASE_URL } from '../src/config/github';
+import { GitHubRepository } from '../src/types/github';
 import githubInterceptor from './interceptors/github';
 
 describe('Example tests', () => {
-  const ownerName = 'owner';
-  const repositoryName = 'example';
-
   const repository: GitHubRepository = {
     id: 1,
-    name: repositoryName,
-    full_name: `${ownerName}/${repositoryName}`,
-    html_url: `https://github.com/${ownerName}/${repositoryName}`,
-    owner: { login: ownerName },
+    name: 'zimic',
+    full_name: 'zimicjs/zimic',
+    html_url: 'https://github.com/zimicjs/zimic',
+    owner: { login: 'zimicjs' },
   };
 
-  beforeEach(() => {
-    document.body.innerHTML = '';
-  });
-
-  it('should render a GitHub repository, if found', async () => {
+  it('should return a GitHub repository, if found', async () => {
     githubInterceptor
-      .get(`/repos/${ownerName}/${repositoryName}`)
+      .get('/repos/zimicjs/zimic')
       .respond({
         status: 200,
         body: repository,
       })
       .times(1);
 
-    renderApp();
+    const response = await fetch(`${GITHUB_API_BASE_URL}/repos/zimicjs/zimic`);
+    expect(response.status).toBe(200);
 
-    const ownerInput = screen.getByRole('textbox', { name: 'Owner' });
-    await userEvent.type(ownerInput, ownerName);
-
-    const repositoryInput = screen.getByRole('textbox', { name: 'Repository' });
-    await userEvent.type(repositoryInput, repositoryName);
-
-    const searchButton = screen.getByRole('button', { name: 'Search' });
-    await userEvent.click(searchButton);
-
-    const repositoryHeading = await screen.findByRole('heading', { name: repository.full_name });
-    expect(repositoryHeading).toBeInTheDocument();
-
-    const repositoryLink = await screen.findByRole('link', { name: repository.html_url });
-    expect(repositoryLink).toBeInTheDocument();
+    const data = (await response.json()) as GitHubRepository;
+    expect(data).toEqual(repository);
   });
 
   it('should return a 404 status code, if the GitHub repository is not found', async () => {
@@ -58,21 +38,10 @@ describe('Example tests', () => {
       })
       .times(1);
 
-    renderApp();
+    const response = await fetch(`${GITHUB_API_BASE_URL}/repos/unknown/unknown`);
+    expect(response.status).toBe(404);
 
-    const ownerInput = screen.getByRole('textbox', { name: 'Owner' });
-    await userEvent.type(ownerInput, ownerName);
-
-    const repositoryInput = screen.getByRole('textbox', { name: 'Repository' });
-    await userEvent.type(repositoryInput, repositoryName);
-
-    const searchButton = screen.getByRole('button', { name: 'Search' });
-    await userEvent.click(searchButton);
-
-    const repositoryHeading = await screen.findByRole('heading', { name: 'Repository not found' });
-    expect(repositoryHeading).toBeInTheDocument();
-
-    const repositoryLinkPromise = screen.findByRole('link', {}, { timeout: 200 });
-    await expect(repositoryLinkPromise).rejects.toThrowError();
+    const data = (await response.json()) as { message: string };
+    expect(data).toEqual({ message: 'Not Found' });
   });
 });
