+++ examples/zimic-interceptor-with-vitest-jsdom/src/config/github.ts
@@ -1,0 +1,1 @@
+export const GITHUB_API_BASE_URL = 'https://api.github.com';
