+++ examples/zimic-interceptor-with-vitest-jsdom/tests/interceptors/github.ts
@@ -1,6 +1,7 @@
 import { createHttpInterceptor } from '@zimic/interceptor/http';
 
-import { GITHUB_API_BASE_URL, GitHubSchema } from '../../src/clients/github';
+import { GITHUB_API_BASE_URL } from '../../src/config/github';
+import { GitHubSchema } from '../../src/types/github';
 
 const githubInterceptor = createHttpInterceptor<GitHubSchema>({
   baseURL: GITHUB_API_BASE_URL,
