+++ examples/zimic-interceptor-with-vitest-node/package.json
@@ -8,17 +8,11 @@
     "types:check": "tsc --noEmit"
   },
   "dependencies": {
-    "@zimic/fetch": "latest",
-    "@zimic/http": "latest",
-    "fastify": "^5.3.0",
-    "zod": "^3.24.2"
+    "@zimic/http": "latest"
   },
   "devDependencies": {
-    "@types/supertest": "^6.0.3",
-    "@vitest/spy": "^3.1.1",
     "@zimic/interceptor": "latest",
     "dotenv-cli": "^8.0.0",
-    "supertest": "^7.1.0",
     "typescript": "^5.8.3",
     "vitest": "^3.1.1"
   }
