+++ examples/zimic-interceptor-with-vitest-node/vitest.config.mts
@@ -1,5 +1,6 @@
 /// <reference types="vitest" />
 
+import path from 'path';
 import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
@@ -8,5 +9,11 @@ export default defineConfig({
     include: ['./tests/**/*.test.ts'],
     setupFiles: ['./tests/setup.ts'],
     clearMocks: true,
+  },
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, './src'),
+      '@tests': path.resolve(__dirname, './tests'),
+    },
   },
 });
